scripts/requirements.txt
# Dependances Python pour la generation de documents
# Installation: pip install -r requirements.txt

# Generation DOCX (obligatoire)
python-docx>=1.1.0

# Conversion Markdown vers HTML (pour PDF)
markdown>=3.5

# Generation PDF (optionnel - choisir une option)
# Option 1: weasyprint (pur Python, bonne qualite)
weasyprint>=60.0

# Dependances de weasyprint
Pillow>=10.0
cssselect2>=0.7.0
tinycss2>=1.2.0

# -------------------------------------------
# ALTERNATIVES POUR PDF (non-Python)
# -------------------------------------------
#
# Option A: md-to-pdf (Node.js) - RECOMMANDE
#   npm install -g md-to-pdf
#
# Option B: Pandoc + LaTeX (professionnel)
#   Windows: winget install JohnMacFarlane.Pandoc
#            winget install MiKTeX.MiKTeX
#
# Option C: wkhtmltopdf
#   Windows: winget install wkhtmltopdf.wkhtmltopdf
# -------------------------------------------
